[ALPS07133003] pwm: fix BUF0_SIZE and BUF1_SIZE to 16bits
1. FIX PWMx_BUF0_SIZE and PWMx_BUF1_SIZE to 16 bits
2. export mt_pwm_clk_sel_hal() for customers

MTK-Commit-Id: 1f5faeefdddf5e4835787f1e147d58fbfe9d09f6

CR-Id: ALPS07133003
Feature: irtx

+++ drivers/misc/mediatek/include/mt-plat/mtk_pwm.h
@@ -59,7 +59,7 @@ struct pwm_spec_config {
 			u16 LDURATION;
 			u16 GDURATION;
 			dma_addr_t BUF0_BASE_ADDR;
-			u32 BUF0_SIZE;
+			u16 BUF0_SIZE;
 			u16 WAVE_NUM;
 		} PWM_MODE_MEMORY_REGS;
 
@@ -72,9 +72,9 @@ struct pwm_spec_config {
 			u16 LDURATION;
 			u16 GDURATION;
 			dma_addr_t BUF0_BASE_ADDR;
-			u32 BUF0_SIZE;
+			u16 BUF0_SIZE;
 			dma_addr_t BUF1_BASE_ADDR;
-			u32 BUF1_SIZE;
+			u16 BUF1_SIZE;
 			u16 WAVE_NUM;
 			u32 VALID;
 		} PWM_MODE_RANDOM_REGS;
